Update.cc : communicate first, update next
In practice, this modification does not change the code's behavior because at the beginning all the spins---including the ghosts---are set to 1, so communicating them has no effect. However, if the initialization were different, then communicating first would ensure that the spin-flip probability at iteration 1 at all interior boundary points is calculated using the updated ghosts.